compat: keep Factory decoder within package file budget
